Visual tweaks
Added the "doom" themes, and use "doom-laserwave" for a default. Also
added a GC bump to help with LSP typescript, and tried to finagle
packages further so that I could get some of the lazy-loading stuff to
retain configuration on load (I'm seeing a weird error where the TS
indent doesn't behave correctly).